FFmpeg 2.0 with OpenCFL support
This build comes totally stripped of the Broadcom HD video decoder
acceleration (libcrystalHD) since the SDK brings issues on build time,
and memory leaks are not a part of good code.

Secondly, most of you who use FFmpeg do not do it on underpowered
systems with Broadcom HD video decoders, like Intel Atom SoCs that come
packed with the chip in many netbook designs.

Back to code and coffee. Enjoy.